# server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Student Knowledge Control System API
  description: API системы контроля знаний студентов
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Локальный сервер
...
paths:
...
  /admin/groups:
    get:
      tags: [Admin Groups]
      summary: Получить активные учебные группы
      operationId: getGroups
      responses:
        '200':
          description: Список групп
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Admin Groups]
      summary: Создать учебную группу
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '201':
          description: Группа создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/groups/{id}:
    delete:
      tags: [Admin Groups]
      summary: Выполнить soft delete группы
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/groups/{groupId}/subjects:
    get:
      tags: [Admin Groups]
      summary: Получить предметы группы
      operationId: getGroupSubjects
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Список предметов группы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Admin Groups]
      summary: Назначить предмет группе
      operationId: addSubjectToGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSubjectRequest'
      responses:
        '201':
          description: Связь предмета с группой создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/students:
    get:
      tags: [Admin Students]
      summary: Получить активных студентов
      operationId: getStudents
      responses:
        '200':
          description: Студенты, сгруппированные по учебным группам
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentsByGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/students/search:
    get:
      tags: [Admin Students]
      summary: Найти студентов для привязки к группе
      operationId: searchStudents
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            default: ''
          description: Поиск по ФИО или email
        - name: excluded_group_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Id'
          description: Исключить студентов, уже состоящих в этой группе
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
          description: Максимум 50 студентов
      responses:
        '200':
          description: Список студентов, ограниченный по limit
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Требуется корректный access-токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Связь или сущность уже существует
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    SuccessMessage:
      description: Операция выполнена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  schemas:
    Id:
      type: integer
      format: int64
      minimum: 1
      example: 1
...
    Student:
      type: object
      required: [id, email, full_name, created_at]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          format: email
        full_name:
          type: string
        created_at:
          type: string
          format: date-time
    StudentsByGroup:
      type: object
      required: [group_id, group_name, students]
      properties:
        group_id:
          oneOf:
            - $ref: '#/components/schemas/Id'
            - type: 'null'
        group_name:
          type: string
          example: Нет группы
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
    Group:
      type: object
      required: [id, name]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
    Subject:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        created_at:
          type: string
          format: date-time
    NameRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    AddSubjectRequest:
      type: object
      additionalProperties: false
      required: [subject_id]
      properties:
        subject_id:
          $ref: '#/components/schemas/Id'
...
    SubjectGroup:
      type: object
      required: [id, subject_id, group_id]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        subject_id:
          $ref: '#/components/schemas/Id'
        group_id:
          $ref: '#/components/schemas/Id'
...
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string